Change pattern id condition to use PatternId instead of int

+++ DjmaxRandomSelectorV/Models/AdvancedFilter.cs
@@ -39,7 +39,7 @@ namespace DjmaxRandomSelectorV.Models
 
         public ICondition ToCondition()
         {
-            var result = PatternList.Where(p => _trackDB.Find(p.TrackId)?.IsPlayable ?? false).Select(p => (int)p.Id);
+            var result = PatternList.Where(p => _trackDB.Find(p.TrackId)?.IsPlayable ?? false).Select(p => p.Id);
             return new PatternIdCondition(result);
         }
     }
